requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
openpyxl>=3.1.0
plotly>=5.19.0
openai>=1.12.0
python-dotenv>=1.0.0
numpy>=1.26.0
